--- data/adverbs
@@ -1,588 +1,0 @@
-highly
-happily
-splendidly
-kookily
-solely
-selfishly
-dearly
-measurably
-knowledgeably
-confusedly
-sharply
-sweetly
-tensely
-spectacularly
-likely
-handsomely
-lastly
-gracefully
-briskly
-obnoxiously
-heavily
-safely
-spasmodically
-readily
-hungrily
-thoroughly
-wisely
-ultimately
-widely
-viciously
-methodically
-idiotically
-justly
-surely
-intensely
-restfully
-badly
-partially
-substantially
-loyally
-hysterically
-daily
-angrily
-haphazardly
-covertly
-fully
-knottily
-lithely
-vaguely
-warmly
-resentfully
-unanimously
-imaginatively
-orderly
-diligently
-positively
-reliably
-yearly
-righteously
-intermittently
-independently
-unfavorably
-correctly
-offensively
-sporadically
-vigorously
-weekly
-temporarily
-noisily
-sadly
-fiery
-icily
-selflessly
-expressly
-blissfully
-humorously
-sincerely
-barely
-playfully
-clearly
-primly
-universally
-properly
-constantly
-freely
-voluntarily
-tightly
-specifically
-admiringly
-majestically
-inappropriately
-favorably
-luckily
-poetically
-normally
-coolly
-affirmatively
-eventually
-officially
-ethically
-rightly
-visually
-agreeably
-obediently
-singularly
-casually
-victoriously
-zestily
-currently
-naughtily
-earnestly
-deliberately
-silently
-elegantly
-painfully
-proudly
-poorly
-furiously
-dreamily
-materially
-nastily
-touchingly
-externally
-possibly
-worriedly
-excitedly
-slyly
-luxuriously
-exceptionally
-vastly
-eternally
-startlingly
-generally
-structurally
-wearily
-habitually
-abnormally
-politely
-principally
-fashionably
-bravely
-fondly
-wonderfully
-vertically
-plainly
-explicitly
-poignantly
-probably
-mysteriously
-fiercely
-merely
-steadily
-valiantly
-customarily
-why
-diagonally
-emotionally
-lately
-unnecessarily
-magically
-nearly
-docilely
-tepidly
-joyfully
-wildly
-ferociously
-carefully
-negatively
-perfectly
-cunningly
-crossly
-lazily
-carelessly
-absolutely
-cleverly
-extensively
-miserably
-simply
-unexpectedly
-impressively
-boyishly
-stupidly
-indubitably
-medically
-informally
-only
-comfortably
-boldly
-queasily
-sternly
-intently
-endlessly
-zestfully
-routinely
-directly
-mainly
-presumably
-harshly
-significantly
-optimistically
-stealthily
-bewilderingly
-generously
-scientifically
-limply
-sufficiently
-rudely
-divisively
-ruthlessly
-recklessly
-cowardly
-hopelessly
-every
-respectfully
-easily
-gratefully
-patiently
-respectably
-lovingly
-curiously
-softly
-reassuringly
-dimly
-today
-brilliantly
-ironically
-shrilly
-famously
-mostly
-monthly
-exclusively
-lightly
-exceedingly
-uselessly
-urgently
-elaborately
-wetly
-bitterly
-unskillfully
-shortly
-paradoxically
-immediately
-instantly
-busily
-innocently
-securely
-mortally
-jauntily
-nightly
-inquisitively
-repeatedly
-markedly
-oddly
-merrily
-fatally
-periodically
-slightly
-enjoyably
-incessantly
-extraordinarily
-secretively
-evocatively
-hauntingly
-still
-unselfishly
-fervently
-successfully
-genuinely
-specially
-rarely
-clumsily
-accordingly
-believably
-early
-considerately
-reasonably
-outrageously
-thereby
-daintily
-anyway
-evenly
-honestly
-mockingly
-actually
-cloudily
-enthusiastically
-daringly
-immensely
-chiefly
-invariably
-suitably
-stressfully
-surprisingly
-roughly
-painlessly
-dastardly
-beautifully
-densely
-recently
-impatiently
-healthily
-quizzically
-powerfully
-frequently
-upwardly
-unbelievably
-particularly
-finally
-bleakly
-evidently
-persistently
-especially
-hastily
-effortlessly
-separately
-smoothly
-dopily
-richly
-jealously
-longingly
-tremendously
-supremely
-physically
-immeasurably
-unnaturally
-unquestionably
-vacantly
-openly
-speedily
-awfully
-infrequently
-sedately
-accidentally
-compactly
-visibly
-scarcely
-hurriedly
-bewilderedly
-unemotionally
-affectionately
-closely
-shamelessly
-woefully
-wrongly
-greatly
-strictly
-deftly
-blindly
-outwardly
-arrogantly
-acutely
-fluently
-stylishly
-obligingly
-skillfully
-frenetically
-previously
-grimly
-keenly
-madly
-naturally
-foolishly
-subtly
-overconfidently
-ridiculously
-appreciably
-usually
-vigilantly
-seemingly
-abundantly
-bashfully
-defiantly
-youthfully
-irritably
-shyly
-financially
-meanly
-internally
-gleefully
-scarily
-quaintly
-searchingly
-loudly
-incorrectly
-utterly
-wiggly
-wickedly
-terribly
-awkwardly
-formerly
-solemnly
-loosely
-jaggedly
-graciously
-silicitiously
-suspiciously
-entirely
-loftily
-unwillingly
-extremely
-meaningfully
-assuredly
-obviously
-presently
-ordinarily
-weakly
-lively
-cleanly
-spiritually
-sleepily
-violently
-charmingly
-courageously
-irrefutably
-annoyingly
-already
-sympathetically
-willingly
-inevitably
-girlishly
-rapidly
-reponsibly
-zanily
-helpfully
-studiously
-calmly
-actively
-unfailingly
-dangerously
-mechanically
-nosily
-compassionately
-yesterday
-enormously
-nicely
-faithfully
-jovially
-consequently
-coincidentally
-frankly
-persuasively
-absentmindedly
-nervously
-doubtfully
-quietly
-potentially
-vainly
-kindheartedly
-prettily
-systematically
-seriously
-thankfully
-essentially
-colorfully
-yearningly
-very
-hatefully
-gently
-needily
-joyously
-handily
-coldly
-morally
-willfully
-unfortunately
-rightfully
-worthily
-nearby
-differently
-darkly
-passionately
-invisibly
-formally
-slowly
-knavishly
-continuously
-tenderly
-suddenly
-wholly
-numbly
-menacingly
-verbally
-truthfully
-reluctantly
-half-heartedly
-truly
-quickly
-considerably
-equally
-boastfully
-candidly
-amazingly
-promptly
-fortunately
-usefully
-neatly
-swiftly
-solidly
-ominously
-rigidly
-remarkably
-horizontally
-really
-punctually
-exactly
-strenuously
-vivaciously
-delicately
-hourly
-subsequently
-regularly
-gradually
-well
-eloquently
-thoughtfully
-perpetually
-efficiently
-questionably
-childishly
-gladly
-jubilantly
-completely
-whole-heartedly
-energetically
-annually
-commonly
-begrudgingly
-sheepishly
-deeply
-kiddingly
-judgmentally
-consistently
-appropriately
-infinitely
-reproachfully
-coaxingly
-kindly
-finitely
-delightfully
-randomly
-heartily
-fairly
-decently
-harmoniously
-brightly
-crazily
-dramatically
-hugely
-puzzlingly
-cautiously
-astonishingly
-certainly
-unkindly
-uniformly
-zealously
-frantically
-anxiously
-briefly
-interestingly
-questioningly
-personally
-unabashedly
-upliftingly
-intelligently
-shakily
-unilaterally
-brutally
-disorderly
-unethically
-mightily
-cruelly
-quirkily
-joylessly
-cheerfully
-grudgingly
-occasionally
-continually
-superficially
-cutely
-identifiably
-unbearably
-unimpressively
-knowingly
-totally
-greedily
-inwardly
-shoddily
-eagerly
