--- data/other
@@ -1,73 +1,0 @@
-aboard
-somewhat
-never
-somehow
-upward
-next
-nowadays
-enough
-late
-evermore
-about
-meantime
-sometimes
-abroad
-always
-anyhow
-tomorrow
-extra
-quite
-seldom
-forever
-less
-meanwhile
-just
-upright
-somewhere
-later
-elsewhere
-when
-before
-thus
-everywhere
-outdoors
-up
-so
-indoors
-furthermore
-forward
-afterwards
-fast
-upbeat
-outwards
-soon
-hence
-now
-indeed
-upside-down
-nevertheless
-hitherto
-around
-yet
-aside
-then
-during
-quicker
-overseas
-much
-while
-almost
-anywhere
-far
-moreover
-there
-anyplace
-however
-once
-down
-where
-twice
-after
-more
-all
-away
