--- data/verbs
@@ -1,910 +1,0 @@
-pines
-tips
-hooks
-alleges
-thunders
-prys
-proceeds
-maintains
-sputters
-knots
-trains
-admires
-brushes
-bargains
-broils
-forces
-chides
-bombs
-requests
-challenges
-promises
-scolds
-raises
-uses
-rolls
-stuns
-spares
-impresses
-hands
-drips
-marks
-roars
-explores
-happens
-stammers
-hovers
-telephones
-echoes
-bats
-remembers
-flosses
-exclaims
-orders
-recognizes
-teases
-bandages
-mixes
-returns
-loses
-snoozes
-ends
-smashes
-laments
-offends
-grins
-strums
-continues
-radiates
-replys
-visits
-shelters
-reflects
-shades
-inspects
-announces
-shaves
-thanks
-explains
-lifts
-releases
-grills
-trys
-itches
-pleases
-nests
-excites
-mugs
-shivers
-votes
-bores
-hops
-extolls
-harms
-yawns
-cures
-proclaims
-stresses
-sees
-alludes
-travels
-quilts
-enters
-trusts
-plays
-stirs
-bumps
-covets
-invests
-fences
-nods
-gives
-describes
-encodes
-repairs
-rushes
-cross-examines
-sells
-jails
-runs
-infuses
-reprimands
-scares
-predicts
-recommends
-punches
-knits
-rinses
-confides
-alerts
-reduces
-trickles
-attaches
-improves
-times
-drowns
-gets
-fishes
-evacuates
-folds
-marvels
-severs
-emits
-borrows
-drags
-relates
-gurgles
-doles
-expands
-flows
-stabs
-doubts
-immigrates
-suits
-shoots
-multiplys
-combs
-prays
-lands
-professes
-receives
-types
-skips
-explodes
-parks
-pedals
-attends
-shuts
-supplys
-punishes
-mars
-feeds
-snows
-shrugs
-rings
-hinders
-succeeds
-loves
-yells
-retorts
-sketches
-begins
-drives
-puts
-urges
-belongs
-jumps
-barters
-scoffs
-gazes
-admits
-delivers
-permits
-boils
-delays
-feels
-suffers
-heaps
-irritates
-greets
-vanquishes
-bruises
-breaks
-reiterates
-dares
-twists
-tows
-searches
-freezes
-implores
-makes
-bastes
-inspires
-assures
-sneaks
-stutters
-suspects
-keeps
-fools
-sparkles
-utters
-frets
-troubles
-shops
-traps
-kills
-mines
-shouts
-faces
-heats
-decays
-dusts
-screeches
-snorts
-bears
-notes
-gushes
-murders
-declares
-claps
-stores
-lists
-delights
-falters
-jogs
-mumbles
-sips
-barrages
-spends
-sprays
-comforts
-decides
-lights
-scratches
-lightens
-pesters
-protects
-branches
-berates
-breathes
-stipulates
-hails
-avoids
-guesses
-enunciates
-vouches
-holds
-breeds
-speaks
-paints
-plans
-highlights
-tells
-avows
-removes
-commands
-hates
-complains
-oils
-supposes
-babbles
-screws
-senses
-embarrasses
-joins
-pushes
-tames
-stops
-gnaws
-believes
-increases
-attracts
-prevents
-craves
-sounds
-dresses
-rides
-sways
-regrets
-prints
-gasps
-listens
-admonishes
-attempts
-brings
-leans
-ogles
-phones
-nags
-rejoices
-marches
-pastes
-stashes
-extends
-starts
-discloses
-teaches
-bites
-handles
-licenses
-cracks
-consists
-bows
-builds
-pads
-dotes
-rises
-fades
-adds
-lies
-confirms
-does
-brakes
-analyzes
-falls
-usurps
-rebuffs
-interferes
-frys
-shampooes
-auctions
-cuts
-reveals
-arrives
-pleads
-deserts
-performs
-hunts
-pledges
-informs
-criticizes
-bares
-grips
-dyes
-stamps
-heals
-boxes
-concerns
-unlocks
-squashes
-infects
-chips
-labels
-speculates
-greases
-naps
-understands
-cautions
-places
-skis
-trips
-sings
-relaxes
-rains
-yields
-mutters
-asks
-picks
-unties
-jams
-means
-hopes
-chokes
-presses
-shares
-fixes
-quotes
-steps
-sympathizes
-colors
-chants
-besieges
-soaks
-snatches
-irons
-suspends
-jabbers
-sends
-scoots
-mentions
-earns
-treats
-contains
-theorizes
-states
-fumes
-claims
-emigrates
-equivocates
-trades
-deters
-spots
-produces
-toasts
-crawls
-agrees
-jokes
-bestows
-pauses
-croaks
-yodels
-sneers
-switches
-coughs
-signals
-lasts
-tempts
-catches
-offers
-singes
-harnesses
-brands
-pets
-drones
-thaws
-copys
-behaves
-jests
-tastes
-drinks
-identifys
-barks
-exists
-affords
-concentrates
-proposes
-signs
-stuffs
-reaches
-concurs
-communicates
-notices
-sows
-apologizes
-hangs
-begs
-operates
-rates
-crashes
-hammers
-counters
-supports
-spills
-hints
-crushes
-swoops
-tours
-plants
-overflows
-glues
-screams
-rants
-punctures
-calls
-vacuums
-lowers
-preaches
-hums
-cleans
-interests
-corrects
-rocks
-invents
-soothes
-blushes
-omits
-bars
-fights
-stitches
-instructs
-protests
-enchants
-ices
-programs
-comments
-forgets
-dumps
-rejects
-volunteers
-escapes
-lives
-sneezes
-emphasizes
-sleeps
-directs
-leaves
-crosses
-doubles
-changes
-mates
-flashes
-jabs
-crowds
-floats
-enjoys
-snickers
-burys
-reminds
-robs
-bubbles
-sets
-bleeds
-sews
-bakes
-pulls
-manages
-surmises
-deals
-envys
-dislikes
-trims
-favors
-questions
-fires
-moves
-dances
-steers
-debates
-tosses
-numbers
-cooks
-pours
-deserves
-occurs
-bounces
-misses
-ignores
-loans
-smiles
-submerges
-ties
-risks
-kneels
-surprises
-dives
-judges
-charges
-names
-learns
-practices
-foretells
-memorizes
-ventures
-bids
-confesses
-sniffs
-levels
-asserts
-realizes
-obtains
-dials
-reads
-draws
-chops
-concedes
-entertains
-gathers
-merges
-hears
-recites
-tests
-demands
-affirms
-faxes
-encourages
-shows
-snarls
-opens
-damages
-employs
-saws
-educates
-contends
-chooses
-collects
-beats
-prepares
-decorates
-files
-boasts
-peers
-zooms
-divulges
-consents
-reports
-costs
-covers
-introduces
-arranges
-mourns
-befriends
-answers
-separates
-packs
-crys
-provides
-brags
-divides
-paddles
-applauds
-bleaches
-jolts
-parts
-grunts
-dupes
-growls
-shifts
-glides
-spins
-leads
-shrieks
-postulates
-transports
-records
-overruns
-rows
-forgives
-cheats
-knows
-sprouts
-pretends
-slides
-bets
-replaces
-burns
-books
-struts
-flaps
-acts
-burnishes
-pants
-exhorts
-squeezes
-obeys
-sparks
-drops
-trembles
-blinks
-likes
-bans
-hits
-traces
-scrapes
-hisses
-curls
-becomes
-fills
-mows
-exaggerates
-inflates
-laughs
-passes
-remains
-blurts
-vanishes
-ticks
-scribbles
-invites
-surrounds
-appears
-appeals
-pelts
-molts
-advises
-frames
-juggles
-matters
-storms
-founds
-meets
-swims
-relys
-connects
-grows
-lends
-stays
-includes
-slips
-suggests
-dies
-upsets
-counts
-adopts
-translates
-posts
-interjects
-curves
-allows
-dreams
-tires
-rots
-cares
-settles
-slaps
-stands
-goes
-raves
-sighs
-sins
-tracks
-finds
-bawls
-clears
-bellows
-reigns
-docks
-preserves
-dictates
-chews
-fetches
-tricks
-testifys
-implys
-interrupts
-examines
-chases
-stunts
-exercises
-saves
-deciphers
-fits
-lectures
-has
-locks
-blinds
-sails
-kicks
-slices
-guarantees
-battles
-drys
-climbs
-snores
-deceives
-frightens
-ails
-values
-rules
-seals
-repeats
-races
-rescues
-expects
-refuses
-snoops
-takes
-grates
-skateboards
-divorces
-taunts
-presents
-throws
-knocks
-attacks
-bobs
-gardens
-imagines
-moors
-digresses
-completes
-balks
-points
-threatens
-thinks
-bolts
-turns
-causes
-beams
-backs
-hides
-spells
-scatters
-broadcasts
-fears
-tickles
-smells
-owns
-coils
-quizs
-talks
-pokes
-flys
-camps
-inquires
-guides
-calculates
-looks
-yelps
-evaporates
-elopes
-pricks
-smokes
-trots
-arrests
-observes
-owes
-blesses
-closes
-tumbles
-mails
-says
-compares
-scorches
-checks
-influences
-haunts
-films
-argues
-zips
-stars
-acknowledges
-considers
-measures
-strays
-coaches
-unites
-mends
-injures
-balances
-precedes
-digs
-melts
-stares
-hurts
-emptys
-voices
-excuses
-chomps
-sobs
-terrifies
-appreciates
-concludes
-forms
-rhymes
-hurrys
-denys
-guards
-approves
-accepts
-gilds
-cycles
-polishes
-follows
-trounces
-ponders
-clips
-spoils
-dams
-unpacks
-reverses
-savors
-drills
-carves
-cheers
-needs
-fails
-buzzs
-quits
-objects
-depends
-upholds
-carrys
-flees
-hijacks
-develops
-expels
-pays
-ruins
-annoys
-spans
-aches
-crochets
-insists
-howls
-recalls
-buys
-hoots
-competes
-fastens
-possesses
-cajoles
-snaps
-shocks
-amuses
-stretches
-retires
-stows
-upstages
-intends
-strengthens
-edits
-strokes
-articulates
-skates
-jeers
-swears
-pats
-swings
-helps
-slows
-grumbles
