--- install-dependencies.sh
@@ -1,9 +1,0 @@
-chicken-install test
-chicken-install combinators
-chicken-install random-bsd
-chicken-install stack
-chicken-install spiffy 
-chicken-install intarweb
-chicken-install medea
-chicken-install spiffy-uri-match
-chicken-install utf8
