+++ lib/index.js
@@ -1,0 +1,7 @@
+var addon = require('../native');
+
+console.log(addon.generate(0, 100));
+console.log("\n");
+console.log(addon.generate(1, 10000));
+
+
