+++ native/Cargo.toml
@@ -21,4 +21,4 @@ rand = "0.5"
 num-traits = "0.2"
 num-derive = "0.3"
 either = "1.5.3"
-unicode-segmentation = "1.6.0"
+lazy_static = "1.4.0"
