+++ native/src/data.rs
@@ -1,0 +1,134 @@
+use std::collections::HashMap;
+use types::*;
+
+macro_rules! map(
+        { $($key:expr => $value:expr),+ } => {
+                {
+                let mut m = ::std::collections::HashMap::new();
+                $(
+                        m.insert($key, $value);
+                )+
+                m
+                }
+        };
+);
+
+lazy_static! {
+        /// The database of words, grouped by grammatical class and then word size
+        pub static ref GRAMATICAL_CLASSES: HashMap<GrammaticalClass, HashMap<u8, Vec<&'static str>>> = {
+                println!("creating classsss");
+                return map! {
+                        GrammaticalClass::Other => map! {
+                                    12u8 => vec!["nevertheless"],
+                                    4u8 => vec!["next", "late", "less", "just", "when", "thus", "fast", "soon", "then", "much", "once", "down", "more", "away"],
+                                    2u8 => vec!["up", "so"],
+                                    7u8 => vec!["somehow", "forever", "upright", "indoors", "forward", "quicker", "however"],
+                                    11u8 => vec!["furthermore", "upside-down"],
+                                    3u8 => vec!["now", "yet", "far", "all"],
+                                    6u8 => vec!["aboard", "upward", "enough", "abroad", "always", "anyhow", "seldom", "before", "upbeat", "indeed", "around", "during", "almost"],
+                                    8u8 => vec!["somewhat", "nowadays", "evermore", "meantime", "tomorrow", "outdoors", "outwards", "hitherto", "overseas", "anywhere", "moreover", "anyplace"],
+                                    10u8 => vec!["everywhere", "afterwards"],
+                                    5u8 => vec!["never", "about", "extra", "quite", "later", "hence", "aside", "while", "there", "where", "twice", "after"],
+                                    9u8 => vec!["sometimes", "meanwhile", "somewhere", "elsewhere"]
+                                    }
+                                    ,GrammaticalClass::Verbs => map! {
+                                    5u8 => vec!["pines", "hooks", "knots", "bombs", "rolls", "stuns", "hands", "drips", "marks", "roars", "mixes", "loses", "grins", "lifts", "nests", "votes", "bores", "harms", "yawns", "cures", "plays", "stirs", "bumps", "gives", "sells", "jails", "knits", "times", "folds", "emits", "drags", "doles", "flows", "stabs", "suits", "combs", "prays", "lands", "types", "skips", "parks", "shuts", "feeds", "snows", "rings", "loves", "yells", "urges", "jumps", "gazes", "boils", "feels", "heaps", "dares", "makes", "keeps", "fools", "frets", "shops", "traps", "kills", "mines", "faces", "heats", "dusts", "bears", "notes", "claps", "lists", "hails", "holds", "plans", "tells", "avows", "hates", "joins", "tames", "stops", "gnaws", "rides", "sways", "gasps", "leans", "ogles", "bites", "dotes", "rises", "fades", "falls", "hunts", "bares", "grips", "heals", "boxes", "chips", "trips", "sings", "rains", "picks", "means", "hopes", "fixes", "steps", "soaks", "irons", "sends", "earns", "fumes", "spots", "jokes", "lasts", "thaws", "copys", "jests", "barks", "signs", "hangs", "rates", "hints", "tours", "glues", "rants", "calls", "rocks", "omits", "dumps", "lives", "mates", "burys", "bakes", "pulls", "deals", "envys", "trims", "fires", "moves", "cooks", "pours", "loans", "risks", "dives", "names", "dials", "reads", "draws", "chops", "hears", "tests", "faxes", "shows", "opens", "beats", "files", "peers", "zooms", "costs", "packs", "brags", "jolts", "parts", "dupes", "spins", "leads", "knows", "burns", "books", "flaps", "pants", "obeys", "drops", "likes", "curls", "fills", "ticks", "pelts", "molts", "meets", "swims", "relys", "grows", "lends", "stays", "slips", "posts", "tires", "cares", "slaps", "raves", "sighs", "finds", "bawls", "docks", "chews", "saves", "locks", "sails", "kicks", "rules", "seals", "races", "takes", "moors", "balks", "bolts", "turns", "beams", "backs", "hides", "fears", "coils", "quizs", "talks", "pokes", "camps", "looks", "yelps", "trots", "mails", "films", "stars", "mends", "melts", "hurts", "forms", "denys", "gilds", "clips", "needs", "fails", "buzzs", "quits", "flees", "ruins", "spans", "aches", "howls", "hoots", "snaps", "stows", "edits", "jeers", "helps", "slows"],
+                                    14u8 => vec!["cross-examines"],
+                                    7u8 => vec!["alleges", "admires", "brushes", "happens", "flosses", "returns", "snoozes", "smashes", "laments", "offends", "pleases", "excites", "shivers", "extolls", "alludes", "travels", "invests", "encodes", "repairs", "infuses", "punches", "reduces", "marvels", "borrows", "relates", "gurgles", "expands", "attends", "supplys", "hinders", "retorts", "belongs", "barters", "permits", "suffers", "bruises", "freezes", "assures", "murders", "falters", "mumbles", "decides", "pesters", "berates", "guesses", "vouches", "removes", "babbles", "dresses", "regrets", "listens", "marches", "stashes", "extends", "teaches", "handles", "rebuffs", "reveals", "arrives", "deserts", "pledges", "informs", "unlocks", "infects", "greases", "relaxes", "mutters", "presses", "jabbers", "bestows", "signals", "catches", "behaves", "affords", "reaches", "concurs", "notices", "crashes", "hammers", "crushes", "screams", "vacuums", "invents", "soothes", "blushes", "forgets", "rejects", "escapes", "sneezes", "directs", "crosses", "doubles", "changes", "flashes", "reminds", "bubbles", "manages", "debates", "numbers", "bounces", "ignores", "charges", "asserts", "obtains", "gathers", "recites", "demands", "affirms", "damages", "employs", "chooses", "reports", "answers", "divides", "paddles", "shrieks", "records", "sprouts", "exhorts", "scrapes", "becomes", "remains", "invites", "appears", "appeals", "advises", "juggles", "matters", "settles", "bellows", "fetches", "battles", "repeats", "rescues", "expects", "refuses", "attacks", "gardens", "tickles", "arrests", "blesses", "tumbles", "coaches", "injures", "excuses", "accepts", "follows", "ponders", "unpacks", "objects", "depends", "upholds", "hijacks", "insists", "recalls", "fastens", "cajoles", "retires", "intends", "strokes"],
+                                    6u8 => vec!["trains", "broils", "forces", "chides", "scolds", "raises", "spares", "hovers", "echoes", "orders", "teases", "strums", "replys", "visits", "shades", "shaves", "thanks", "grills", "itches", "quilts", "enters", "trusts", "covets", "fences", "rushes", "scares", "rinses", "alerts", "drowns", "fishes", "severs", "doubts", "shoots", "pedals", "shrugs", "begins", "drives", "scoffs", "admits", "delays", "greets", "breaks", "twists", "bastes", "sneaks", "utters", "shouts", "decays", "snorts", "gushes", "stores", "spends", "sprays", "lights", "avoids", "breeds", "speaks", "paints", "screws", "senses", "pushes", "craves", "sounds", "prints", "brings", "phones", "pastes", "starts", "cracks", "builds", "brakes", "usurps", "pleads", "stamps", "labels", "places", "yields", "unties", "chokes", "shares", "quotes", "colors", "chants", "scoots", "treats", "states", "claims", "trades", "deters", "toasts", "crawls", "agrees", "pauses", "croaks", "yodels", "sneers", "coughs", "tempts", "offers", "singes", "brands", "drones", "tastes", "drinks", "exists", "stuffs", "spills", "swoops", "plants", "lowers", "cleans", "fights", "sleeps", "leaves", "crowds", "floats", "enjoys", "bleeds", "favors", "dances", "steers", "tosses", "occurs", "misses", "smiles", "kneels", "judges", "learns", "sniffs", "levels", "merges", "snarls", "boasts", "covers", "mourns", "grunts", "growls", "shifts", "glides", "cheats", "slides", "struts", "sparks", "blinks", "traces", "hisses", "laughs", "passes", "blurts", "frames", "storms", "founds", "upsets", "counts", "adopts", "curves", "allows", "dreams", "stands", "tracks", "clears", "reigns", "tricks", "implys", "chases", "stunts", "blinds", "slices", "climbs", "snores", "values", "snoops", "grates", "taunts", "throws", "knocks", "points", "thinks", "causes", "spells", "smells", "guides", "elopes", "pricks", "smokes", "closes", "checks", "haunts", "argues", "strays", "unites", "stares", "emptys", "voices", "chomps", "rhymes", "hurrys", "guards", "cycles", "spoils", "savors", "drills", "carves", "cheers", "carrys", "expels", "annoys", "shocks", "amuses", "skates", "swears", "swings"],
+                                    11u8 => vec!["embarrasses", "understands", "sympathizes", "equivocates", "exaggerates", "skateboards", "appreciates", "strengthens", "articulates"],
+                                    4u8 => vec!["tips", "prys", "uses", "bats", "ends", "trys", "mugs", "hops", "sees", "nods", "runs", "gets", "mars", "puts", "tows", "jogs", "sips", "oils", "nags", "bows", "pads", "adds", "lies", "does", "frys", "cuts", "dyes", "naps", "skis", "asks", "jams", "pets", "sows", "begs", "hums", "bars", "ices", "jabs", "robs", "sets", "sews", "ties", "bids", "saws", "crys", "rows", "bets", "acts", "bans", "hits", "mows", "dies", "rots", "goes", "sins", "fits", "drys", "ails", "bobs", "owns", "flys", "owes", "says", "zips", "digs", "sobs", "dams", "pays", "buys", "pats"],
+                                    12u8 => vec!["concentrates", "communicates", "acknowledges"],
+                                    3u8 => vec!["has"],
+                                    8u8 => vec!["thunders", "proceeds", "sputters", "bargains", "requests", "promises", "explores", "stammers", "exclaims", "bandages", "radiates", "shelters", "reflects", "inspects", "explains", "releases", "stresses", "predicts", "confides", "trickles", "attaches", "improves", "receives", "explodes", "punishes", "succeeds", "sketches", "delivers", "searches", "implores", "inspires", "stutters", "suspects", "sparkles", "troubles", "declares", "delights", "barrages", "comforts", "lightens", "protects", "branches", "breathes", "commands", "supposes", "believes", "attracts", "prevents", "attempts", "rejoices", "licenses", "consists", "confirms", "analyzes", "auctions", "performs", "concerns", "squashes", "cautions", "besieges", "snatches", "suspends", "mentions", "contains", "produces", "switches", "proposes", "operates", "counters", "supports", "preaches", "corrects", "stitches", "protests", "enchants", "programs", "comments", "snickers", "surmises", "dislikes", "deserves", "ventures", "realizes", "concedes", "educates", "contends", "collects", "prepares", "divulges", "consents", "arranges", "provides", "applauds", "bleaches", "overruns", "forgives", "pretends", "replaces", "squeezes", "trembles", "inflates", "vanishes", "connects", "includes", "suggests", "dictates", "testifys", "examines", "lectures", "deceives", "divorces", "presents", "imagines", "scatters", "inquires", "observes", "compares", "scorches", "measures", "balances", "precedes", "approves", "polishes", "trounces", "reverses", "develops", "crochets", "competes", "upstages", "grumbles"],
+                                    9u8 => vec!["maintains", "impresses", "remembers", "continues", "announces", "proclaims", "describes", "evacuates", "multiplys", "professes", "irritates", "screeches", "scratches", "complains", "increases", "discloses", "shampooes", "theorizes", "emigrates", "harnesses", "identifys", "overflows", "punctures", "interests", "instructs", "questions", "submerges", "surprises", "practices", "foretells", "memorizes", "confesses", "decorates", "befriends", "separates", "burnishes", "scribbles", "surrounds", "preserves", "exercises", "deciphers", "frightens", "digresses", "completes", "threatens", "considers", "terrifies", "concludes", "possesses", "stretches"],
+                                    10u8 => vec!["challenges", "telephones", "recognizes", "reprimands", "recommends", "immigrates", "vanquishes", "reiterates", "stipulates", "enunciates", "highlights", "admonishes", "interferes", "criticizes", "speculates", "apologizes", "volunteers", "emphasizes", "entertains", "encourages", "introduces", "postulates", "transports", "translates", "interjects", "interrupts", "guarantees", "broadcasts", "calculates", "evaporates", "influences"]
+                                    }
+                                    ,GrammaticalClass::Nouns => map! {
+                                    13u8 => vec!["komodo dragon", "brew director", "water buffalo", "spirit healer", "chorus singer", "postal worker", "elephant seal", "advertisement", "cryptographer", "mathematician"],
+                                    17u8 => vec!["radio broadcaster"],
+                                    14u8 => vec!["representative", "businessperson", "press operator", "security guard", "street sweeper"],
+                                    16u8 => vec!["flight attendant"],
+                                    9u8 => vec!["structure", "alligator", "cassowary", "deliverer", "afternoon", "passenger", "appliance", "conductor", "assistant", "agreement", "goldfinch", "acoustics", "Englander", "invention", "doorkeepe", "breakfast", "goosander", "rainstorm", "butterfly", "homemaker", "astronaut", "treatment", "underwear", "existence", "secretary", "albatross", "carpenter", "sandpiper", "operation", "armadillo", "red panda", "selection", "expansion", "apparatus", "furniture", "jellyfish", "cormorant", "transport", "substance", "porcupine", "direction", "quicksand", "aftermath", "insurance", "authority", "scapegoat", "territory", "amusement", "crocodile", "attention", "dragonfly", "partridge", "go-getter", "digestion", "wolverine", "cockroach", "enchanter", "lunchroom", "education", "barracuda", "discovery", "vegetable", "knowledge", "condition", "scarecrow", "committee", "statement", "principal", "pollution", "ecologist", "economist", "executive", "caregiver", "chauffeur", "concierge", "custodian", "messenger", "scientist", "violinist", "librarian", "zookeeper", "zoologist"],
+                                    6u8 => vec!["porter", "battle", "flight", "curlew", "galley", "drawer", "rabbit", "sonsun", "bubble", "friend", "sister", "cattle", "tongue", "weight", "cheese", "pencil", "pigeon", "sponge", "dinner", "attack", "bottle", "walrus", "parrot", "coyote", "wealth", "sneeze", "potato", "quartz", "crayon", "eggnog", "locket", "riddle", "cobweb", "quelea", "basket", "turkey", "engine", "stitch", "thread", "memory", "effect", "square", "vessel", "street", "recess", "silver", "icicle", "profit", "mitten", "income", "minute", "marble", "credit", "person", "button", "nation", "motion", "oyster", "amount", "camera", "canvas", "poison", "throat", "expert", "quince", "church", "sailor", "zipper", "banana", "anchor", "thrill", "ferret", "lawyer", "magpie", "weasel", "finger", "flower", "salmon", "clover", "cactus", "harbor", "gerbil", "chance", "falcon", "pocket", "summer", "polish", "kettle", "degree", "reason", "turtle", "alpaca", "bushes", "doctor", "temper", "breath", "faucet", "system", "beetle", "advice", "ticket", "locust", "cleric", "letter", "health", "answer", "border", "beaver", "hornet", "voyage", "garden", "school", "theory", "parcel", "muscle", "cellar", "guitar", "orange", "cherry", "runner", "middle", "baboon", "throne", "singer", "prison", "hammer", "monkey", "pickle", "cannon", "number", "farmer", "chough", "grouse", "beggar", "plough", "flavor", "animal", "powder", "lumber", "needle", "branch", "change", "wombat", "jaguar", "dunlin", "kitten", "spider", "celery", "wrench", "parent", "belief", "design", "insect", "action", "rhythm", "copper", "stream", "office", "window", "detail", "market", "zephyr", "record", "winter", "circle", "regret", "priest", "butter", "badger", "dugong", "jackal", "spring", "police", "collar", "quiver", "bucket", "writer", "string", "liquid", "island", "reward", "bridge", "growth", "ground", "editor", "consul", "marine", "archer", "artist", "author"],
+                                    4u8 => vec!["pump", "club", "swan", "unit", "fire", "yoke", "trip", "beef", "bead", "cast", "duck", "mice", "iron", "sand", "rail", "crib", "rock", "soap", "side", "bird", "hope", "drug", "look", "snow", "sack", "talk", "idea", "card", "lamp", "gaur", "love", "goat", "swim", "lock", "drop", "body", "comb", "nest", "wine", "gold", "pony", "drum", "pipe", "meal", "debt", "slip", "lead", "back", "mine", "heat", "hare", "pail", "skin", "song", "pest", "coil", "stew", "dirt", "face", "sign", "home", "chin", "size", "yarn", "wood", "heir", "bulb", "push", "frog", "whip", "fact", "knee", "mole", "wish", "move", "root", "bean", "dust", "self", "base", "hole", "moon", "wind", "newt", "glue", "kudu", "baby", "game", "room", "bear", "seed", "fall", "mask", "fear", "joke", "view", "tent", "kite", "male", "tray", "wool", "fold", "vase", "boot", "play", "sofa", "wave", "hill", "work", "head", "pain", "team", "bomb", "list", "word", "lion", "twig", "fang", "cook", "beam", "edge", "hall", "cart", "knot", "soup", "crab", "mass", "stem", "copy", "mint", "soda", "nail", "sink", "mink", "fish", "name", "ship", "lace", "vest", "bite", "hour", "wash", "plot", "wing", "sock", "hose", "zinc", "wall", "veil", "coal", "town", "cake", "stop", "ibex", "loaf", "deer", "hook", "wire", "burn", "join", "form", "seat", "star", "rook", "turn", "maid", "wren", "lake", "bone", "step", "rule", "mist", "line", "desk", "rest", "week", "corn", "gull", "rose", "shoe", "rice", "door", "road", "cord", "jail", "hate", "feet", "doll", "pear", "dime", "hero", "cork", "salt", "foot", "dove", "book", "food", "arch", "crow", "boss", "spot", "silk", "tree", "cent", "sail", "land", "time", "nose", "gnat", "rate", "jump", "ring", "army", "pull", "bath", "fork", "lark", "oven", "ball", "park", "tail", "dock", "walk", "help", "suit", "need", "camp", "rain", "shop", "vein", "loss", "fuel", "mark", "clam", "part", "horn", "news", "bike", "coat", "boat", "boar", "rake", "milk", "toad", "wolf", "mind", "meat", "worm", "leaf", "show", "able", "year", "lift", "bait", "tank", "gate", "farm", "roof", "seal", "bell", "neck", "kick", "flag", "page", "yard", "hawk", "fowl", "test", "wasp", "note", "care", "band", "hair", "sort", "ibis", "roll", "cave", "chef", "monk"],
+                                    5u8 => vec!["thumb", "actor", "print", "tiger", "shark", "diver", "quiet", "title", "glass", "apple", "metal", "taste", "finch", "quail", "smile", "crate", "flesh", "waves", "slope", "paper", "waste", "crush", "fifth", "quill", "spoon", "humor", "music", "stamp", "honey", "peace", "drink", "lemur", "jeans", "class", "lunch", "crime", "fight", "trail", "usher", "feast", "scarf", "smoke", "kitty", "otter", "night", "sleep", "linen", "scene", "robot", "chair", "frame", "wound", "magic", "sloth", "river", "error", "color", "goose", "crook", "house", "horse", "laugh", "spark", "geese", "screw", "group", "crowd", "cable", "straw", "price", "wheel", "tramp", "knife", "crack", "teeth", "grade", "order", "water", "sense", "crown", "chess", "birth", "shape", "koala", "brain", "party", "cause", "front", "cover", "burst", "scent", "plane", "plate", "range", "paint", "death", "guide", "floor", "watch", "table", "okapi", "blade", "stone", "whale", "power", "place", "route", "steel", "truck", "month", "sleet", "patch", "drain", "hippo", "train", "flock", "juror", "brush", "limit", "coati", "eagle", "board", "cobra", "shake", "heart", "event", "skate", "voice", "space", "spade", "match", "north", "brass", "field", "skunk", "snail", "glove", "juice", "panda", "money", "basin", "llama", "light", "plant", "thing", "clock", "fruit", "sheet", "anger", "slave", "stage", "owner", "moose", "zebra", "start", "cloth", "nerve", "beast", "louse", "bison", "noise", "cloud", "mouth", "jelly", "coast", "bread", "coach", "camel", "tapir", "cream", "hyena", "brick", "dress", "fixer", "shame", "cough", "shelf", "chalk", "earth", "eland", "shade", "shock", "trick", "alley", "grass", "store", "staff", "shrew", "skirt", "heron", "level", "flame", "shirt", "stove", "story", "blood", "sound", "force", "scale", "robin", "wrist", "steam", "chain", "mouse", "point", "berry", "raven", "prose", "twist", "ghost", "quilt", "snake", "sugar", "tooth", "badge", "brake", "judge", "smell", "queen", "verse", "trade", "rifle", "crane", "sheep", "grape", "loris", "elbow", "curve", "grain", "swing", "jewel", "value", "alarm", "paste", "offer", "angle", "touch", "ocean", "clerk", "mason", "mayor", "miner", "model", "scout", "smith", "valet", "vicar"],
+                                    8u8 => vec!["stranger", "question", "mandrill", "boundary", "learning", "good-bye", "kinkajou", "forester", "creature", "calendar", "dinosaur", "starling", "argument", "business", "relation", "industry", "cemetery", "birthday", "friction", "commoner", "surprise", "ancestor", "mosquito", "downtown", "lyrebird", "sorcerer", "trousers", "mountain", "airplane", "reporter", "vacation", "achiever", "activity", "comedian", "governor", "increase", "notebook", "squirrel", "children", "newcomer", "building", "seahorse", "behavior", "religion", "approval", "carriage", "teaching", "tendency", "umbrella", "sculptor", "hedgehog", "decision", "graduate", "aardvark", "beginner", "anteater", "porpoise", "addition", "pleasure", "enlistee", "cherries", "division", "position", "ornament", "mongoose", "flamingo", "antelope", "stocking", "kangaroo", "seashore", "reaction", "goldfish", "minister", "sidewalk", "scissors", "exchange", "distance", "sea lion", "pheasant", "dotterel", "reindeer", "red deer", "property", "hospital", "daughter", "interest", "baseball", "language", "elephant", "educator", "engineer", "explorer", "exporter", "chaplain", "composer", "magician", "marketer", "mechanic", "muralist", "musician", "surveyor", "animator", "attorney"],
+                                    10u8 => vec!["toothbrush", "comparison", "toothpaste", "salamander", "polar bear", "plantation", "volleyball", "playground", "experience", "fellowship", "chinchilla", "arithmetic", "suggestion", "earthquake", "basketball", "calculator", "sea urchin", "punishment", "attraction", "connection", "discussion", "instrument", "adjustment", "guinea pig", "compatriot", "chimpanzee", "government", "wilderness", "cab driver", "cartoonist", "contractor", "ambassador", "astronomer", "auctioneer", "undertaker"],
+                                    11u8 => vec!["observation", "housekeeper", "floorwalker", "development", "guinea fowl", "caterpillar", "nightingale", "firefighter", "prairie dog", "hummingbird", "stegosaurus", "competition", "flesh eater", "grasshopper", "destruction", "sand dollar", "giant panda", "electrician", "entertainer", "salesperson", "saxophonist", "scuba diver", "stockbroker", "xylophonist"],
+                                    12u8 => vec!["choir member", "distribution", "faith healer", "spokesperson", "superstition", "afterthought", "sharpshooter", "organization", "entomologist", "entrepreneur", "exterminator", "calligrapher", "cardiologist", "cartographer", "mail carrier", "manufacturer", "statistician", "veterinarian"],
+                                    3u8 => vec!["box", "god", "wax", "yak", "tax", "cub", "nut", "pen", "ice", "net", "can", "art", "ray", "cod", "zoo", "ant", "gnu", "cap", "pig", "rub", "rat", "van", "jar", "hot", "emu", "arm", "fan", "key", "dog", "spy", "gun", "lip", "ape", "bed", "ram", "pot", "ear", "cup", "leg", "war", "rod", "run", "air", "egg", "map", "ink", "day", "pet", "sky", "eel", "eye", "use", "toy", "act", "tub", "sea", "top", "hat", "tin", "fog", "jam", "low", "elk", "bat", "owl", "jay", "car", "pan", "cry", "hen", "sun", "pie", "way", "cat", "pin", "yam", "toe", "bun", "bee", "oil", "bit", "end", "fly", "fox", "one", "cop"],
+                                    7u8 => vec!["termite", "buffalo", "thought", "produce", "popcorn", "dogfish", "manatee", "dolphin", "chicken", "process", "vicuña", "request", "panther", "opinion", "feeling", "lookout", "control", "payment", "insurer", "opossum", "partner", "contact", "library", "raccoon", "pelican", "sweater", "cabbage", "aviator", "lapwing", "oatmeal", "narwhal", "reading", "mailbox", "ostrich", "officer", "hydrant", "disease", "example", "history", "hamster", "airport", "student", "kouprey", "quarter", "visitor", "disgust", "purpose", "stretch", "guanaco", "haircut", "creator", "driving", "gorilla", "whistle", "attempt", "goshawk", "lettuce", "manager", "curtain", "caribou", "cleaner", "picture", "railway", "gazelle", "harmony", "octopus", "fiction", "tarsier", "holiday", "account", "achieve", "echidna", "volcano", "mammoth", "current", "sardine", "thunder", "pancake", "brother", "country", "respect", "impulse", "morning", "peafowl", "giraffe", "leopard", "caption", "trouble", "balance", "shooter", "lobster", "wallaby", "apparel", "stomach", "balloon", "comfort", "cushion", "company", "penguin", "hearing", "meeting", "bedroom", "flagger", "fighter", "cracker", "feather", "artisan", "station", "science", "hobbies", "measure", "channel", "receipt", "protest", "writing", "support", "soldier", "cheetah", "texture", "herring", "journey", "servant", "cowhand", "laborer", "machine", "society", "mallard", "believe", "scholar", "weather", "leather", "plastic", "insider", "ladybug", "quetzal", "empress", "emperor", "captain", "cashier", "catcher", "caterer", "cellist", "chemist", "cobbler", "coroner", "courier", "samurai", "senator", "sheriff", "surgeon", "swimmer", "veteran", "athlete"]
+                                    }
+                                    ,GrammaticalClass::Adjectives => map! {
+                                    8u8 => vec!["grateful", "handsome", "strident", "vigilant", "thorough", "splendid", "grounded", "attached", "academic", "adorable", "metallic", "granular", "trusting", "exciting", "possible", "ordinary", "nautical", "gruesome", "thankful", "creative", "absolute", "shocking", "well-lit", "colossal", "jubilant", "shameful", "innocent", "gracious", "idolized", "charming", "confused", "alarming", "artistic", "composed", "anchored", "unsteady", "lustrous", "majestic", "decisive", "precious", "suburban", "informal", "distinct", "familiar", "advanced", "scratchy", "haunting", "cheerful", "critical", "gleaming", "agitated", "careless", "squiggly", "impolite", "cultured", "fabulous", "indolent", "finished", "writhing", "striking", "fearless", "trifling", "enormous", "harmless", "punctual", "terrible", "clueless", "euphoric", "favorite", "fatherly", "educated", "positive", "obedient", "ironclad", "friendly", "gigantic", "vigorous", "vengeful", "reckless", "carefree", "truthful", "terrific", "negative", "horrible", "quixotic", "tempting", "infamous", "well-off", "inferior", "discrete", "bustling", "tattered", "definite", "wretched", "straight", "motherly", "blushing", "circular", "detailed", "pleasant", "pristine", "gripping", "studious", "intrepid", "lopsided", "physical", "whopping", "reliable", "precious", "internal", "infinite", "growling", "pleasing", "uncommon", "spotless", "constant", "variable", "standard", "ecstatic", "sardonic", "flawless", "fruitful", "deserted", "untimely", "sociable", "puzzling", "snarling", "dazzling", "forsaken", "helpless", "generous", "unwieldy", "concrete", "valuable", "animated", "original", "utilized", "blissful", "mindless", "grizzled", "mediocre", "outlying", "polished", "cautious", "required", "handmade", "aromatic", "watchful", "gullible", "unfolded", "glorious", "rotating", "specific", "esteemed", "stunning", "relieved", "wrathful", "peaceful", "faithful", "menacing", "ignorant", "spiteful", "accurate", "crushing", "troubled", "complete", "perfumed", "personal", "graceful", "wasteful", "powerful", "colorful", "criminal", "separate", "periodic", "heavenly", "disloyal", "slippery", "skeletal", "electric", "probable", "parallel", "luminous", "spirited", "previous", "piercing", "frequent", "sizzling", "outgoing", "superior", "dramatic", "scornful", "flippant", "adorable", "unlawful", "tangible", "diligent", "fragrant", "youthful", "juvenile", "forceful", "ultimate", "fumbling", "athletic", "official", "gorgeous", "insecure", "virtuous"],
+                                    9u8 => vec!["scholarly", "delicious", "indelible", "misguided", "excellent", "unwilling", "far-flung", "glamorous", "wonderful", "insidious", "unhealthy", "impartial", "wide-eyed", "pointless", "concerned", "venerated", "identical", "necessary", "brilliant", "unnatural", "yellowish", "practical", "humongous", "satisfied", "mortified", "unsightly", "insistent", "qualified", "evergreen", "long-term", "lumbering", "memorable", "imperfect", "worrisome", "distorted", "political", "well-made", "agreeable", "difficult", "unequaled", "competent", "agonizing", "important", "realistic", "monstrous", "traumatic", "sparkling", "disguised", "rewarding", "neglected", "flustered", "luxurious", "miserable", "whispered", "unwelcome", "overjoyed", "dimwitted", "fortunate", "marvelous", "talkative", "foolhardy", "repentant", "downright", "surprised", "querulous", "firsthand", "breakable", "bewitched", "imaginary", "fantastic", "essential", "excitable", "shameless", "turbulent", "acrobatic", "exhausted", "deafening", "dishonest", "unwritten", "favorable", "vivacious", "dependent", "ambitious", "energetic", "admirable", "cavernous", "heartfelt", "offensive", "unwitting", "delirious", "defensive", "emotional", "expensive", "quizzical", "miniature", "clear-cut", "enchanted", "nocturnal", "abandoned", "ill-fated", "demanding", "spherical", "deficient", "acclaimed", "palatable", "judicious", "beautiful", "powerless", "elaborate", "sarcastic", "grandiose", "well-worn", "downright", "infantile", "conscious", "sorrowful", "sniveling", "different", "authentic", "whirlwind", "honorable", "cluttered", "pertinent", "quarterly", "grotesque", "revolving", "loathsome", "immediate", "frivolous", "repulsive", "automatic", "treasured", "alienated", "courteous", "unselfish", "hilarious", "anguished", "plaintive", "whimsical", "exemplary", "worthless", "secondary", "colorless", "bountiful", "attentive", "dangerous"],
+                                    5u8 => vec!["nippy", "minor", "musty", "happy", "round", "rowdy", "muddy", "ideal", "blank", "white", "fancy", "smart", "noisy", "wrong", "unfit", "quick", "major", "joint", "blind", "rural", "regal", "dense", "corny", "early", "frail", "brisk", "obese", "every", "nifty", "fixed", "clear", "weird", "sandy", "heavy", "lumpy", "flaky", "sharp", "scaly", "scary", "crazy", "thick", "harsh", "noted", "salty", "light", "plush", "meaty", "sweet", "black", "chief", "grave", "wordy", "jumpy", "super", "cheap", "muted", "sunny", "curly", "ample", "soupy", "adept", "fussy", "tense", "young", "equal", "blond", "stale", "close", "these", "large", "loyal", "naive", "rigid", "pushy", "juicy", "soggy", "agile", "husky", "proud", "jolly", "front", "novel", "pesky", "mushy", "curvy", "third", "tepid", "legal", "bumpy", "vague", "fuzzy", "tubby", "plump", "funny", "short", "milky", "other", "bogus", "prime", "stark", "upset", "quiet", "grown", "witty", "mealy", "valid", "brief", "brown", "steep", "those", "which", "misty", "brave", "bulky", "total", "stiff", "whole", "petty", "merry", "hefty", "rusty", "kooky", "baggy", "known", "fluid", "vivid", "tired", "clean", "hasty", "small", "fresh", "phony", "worst", "lucky", "green", "plain", "urban", "staid", "prize", "runny", "vapid", "royal", "awful", "murky", "silly", "bland", "leafy", "alive", "frank", "right", "rough", "weary", "tough", "cruel", "messy", "spicy", "livid", "great", "eager", "bowed", "lanky", "weepy", "loose", "grimy", "shiny", "itchy", "false", "ready", "hairy", "ruddy", "angry", "jaded", "jumbo", "empty", "crisp", "fatal", "bossy", "moral", "first", "giddy", "woozy", "rapid", "moist", "faint", "dizzy", "steel", "glass", "aware", "minty", "dopey", "gross", "yummy", "perky", "zesty", "grand", "lined", "gummy", "tight", "slimy", "vital", "solid", "silky", "basic", "nutty", "bleak", "burly", "swift", "tasty", "handy", "worse", "alert", "testy", "windy", "dirty", "utter", "giant", "needy", "showy", "mixed", "shady"],
+                                    6u8 => vec!["closed", "somber", "hoarse", "flimsy", "annual", "giving", "medium", "humble", "feisty", "poised", "zigzag", "intent", "ringed", "amused", "common", "feline", "coarse", "expert", "clumsy", "square", "bitter", "lively", "creepy", "warped", "pretty", "subtle", "sweaty", "nimble", "fickle", "strict", "shoddy", "pastel", "unripe", "dreary", "untidy", "stormy", "barren", "famous", "stable", "homely", "severe", "frosty", "putrid", "untrue", "hearty", "sneaky", "shabby", "frugal", "oblong", "greedy", "edible", "direct", "hidden", "acidic", "silver", "sugary", "sudden", "usable", "mellow", "cloudy", "wilted", "silent", "female", "lonely", "actual", "tragic", "bronze", "unruly", "entire", "fluffy", "loving", "grumpy", "liquid", "oblong", "modest", "paltry", "droopy", "slushy", "creamy", "secret", "decent", "wobbly", "quirky", "measly", "thorny", "hollow", "lavish", "klutzy", "queasy", "inborn", "stingy", "guilty", "bouncy", "upbeat", "strong", "unsung", "knobby", "normal", "linear", "gentle", "hungry", "sleepy", "yellow", "smooth", "gloomy", "boring", "frayed", "pricey", "polite", "french", "formal", "robust", "watery", "uneven", "joyful", "second", "little", "trusty", "useful", "flawed", "frizzy", "wiggly", "ragged", "joyous", "kindly", "jaunty", "dental", "superb", "yearly", "costly", "winged", "drafty", "tender", "gifted", "cheery", "forked", "dapper", "dreary", "timely", "wooden", "frozen", "mature", "crafty", "single", "stupid", "golden", "lawful", "flashy", "glossy", "aching", "doting", "frilly", "snappy", "wicked", "sinful", "kosher", "speedy", "dismal", "elated", "exotic", "modern", "portly", "chubby", "starry", "unique", "united", "bright", "impish", "jagged", "honest", "remote", "quaint", "knotty", "narrow", "smoggy", "sparse", "impure", "tricky", "chilly", "active", "skinny", "proper", "afraid", "rotten", "worthy", "cooked", "scarce", "violet", "sticky", "scared", "filthy", "purple", "arctic", "likely", "clever", "shrill", "canine", "weekly", "orange", "better", "lovely", "webbed", "spiffy", "woeful", "double", "candid", "dismal", "adored", "unused", "jovial", "ornery", "cuddly", "svelte", "caring", "frigid", "broken", "junior", "sturdy", "ornate", "serene", "tinted", "deadly", "bubbly", "recent", "grubby", "slight", "meager", "snoopy", "vacant", "simple", "daring"],
+                                    4u8 => vec!["left", "lost", "torn", "damp", "keen", "fake", "soft", "blue", "icky", "weak", "huge", "rare", "wide", "puny", "tart", "firm", "loud", "cool", "numb", "vain", "nice", "near", "good", "calm", "rash", "fast", "next", "posh", "used", "poor", "bold", "lone", "smug", "tidy", "cold", "wild", "lame", "kind", "each", "neat", "wavy", "bare", "sane", "live", "past", "sore", "evil", "fond", "ripe", "pure", "bony", "worn", "drab", "fine", "wary", "taut", "tiny", "that", "glum", "oval", "lean", "flat", "some", "cute", "rich", "late", "spry", "slow", "open", "only", "wise", "pink", "able", "back", "zany", "dear", "full", "thin", "oily", "pale", "long", "sick", "real", "hard", "dual", "ugly", "deep", "this", "ajar", "sour", "safe", "even", "aged", "best", "vast", "both", "trim", "easy", "last", "dark", "high", "warm", "mild", "tall", "slim", "twin", "rude", "free", "limp", "busy", "true", "lazy", "male", "arid", "same", "dull", "half", "wiry", "grim", "meek", "gray", "fair", "dead", "idle", "rosy", "tame", "mean"],
+                                    11u8 => vec!["sure-footed", "stimulating", "enlightened", "cylindrical", "magnificent", "prestigious", "comfortable", "embarrassed", "extroverted", "defenseless", "embellished", "illustrious", "impractical", "aggravating", "extra-small", "kindhearted", "considerate", "understated", "extra-large", "pessimistic", "frightening", "big-hearted", "trustworthy", "responsible", "experienced", "mountainous", "unimportant", "cooperative", "sympathetic", "warmhearted", "coordinated", "outstanding", "unfortunate", "neighboring", "complicated", "imaginative", "adventurous", "intelligent", "substantial", "astonishing", "interesting", "quarrelsome", "spectacular", "appropriate", "superficial", "impassioned", "sentimental", "waterlogged", "unconscious", "descriptive", "rectangular", "intentional", "instructive", "calculating", "everlasting", "unrealistic", "second-hand"],
+                                    12u8 => vec!["multicolored", "incomparable", "good-natured", "questionable", "self-assured", "apprehensive", "accomplished", "ill-informed", "self-reliant", "conventional", "incompatible", "lighthearted", "unacceptable", "quick-witted", "affectionate", "hard-to-find", "well-groomed"],
+                                    13u8 => vec!["sophisticated", "insubstantial", "insignificant", "kaleidoscopic", "inexperienced", "knowledgeable", "irresponsible", "compassionate", "international", "imperturbable", "old-fashioned", "biodegradable", "uncomfortable", "well-informed"],
+                                    14u8 => vec!["quintessential", "happy-go-lucky", "impressionable"],
+                                    3u8 => vec!["all", "fat", "red", "tan", "dim", "wee", "new", "big", "key", "few", "shy", "far", "bad", "ill", "wry", "icy", "apt", "our", "dry", "mad", "sad", "any", "raw", "low", "wet", "hot", "wan", "odd", "old"],
+                                    7u8 => vec!["unknown", "eminent", "welcome", "foolish", "optimal", "limited", "likable", "scrawny", "pitiful", "spotted", "angelic", "crooked", "yawning", "pungent", "jealous", "anxious", "massive", "damaged", "decimal", "mundane", "puzzled", "pointed", "mammoth", "devoted", "admired", "ashamed", "subdued", "limping", "enraged", "amusing", "thirsty", "austere", "soulful", "belated", "scented", "envious", "genuine", "untried", "striped", "unaware", "complex", "natural", "delayed", "antique", "unlined", "stained", "buzzing", "overdue", "prudent", "radiant", "helpful", "unusual", "fitting", "gaseous", "violent", "unlucky", "roasted", "stylish", "assured", "lovable", "regular", "knowing", "rubbery", "babyish", "immense", "profuse", "present", "potable", "healthy", "shallow", "careful", "plastic", "trained", "growing", "popular", "gleeful", "warlike", "leading", "alarmed", "fearful", "failing", "focused", "wealthy", "miserly", "noxious", "jittery", "elderly", "idiotic", "nervous", "glaring", "excited", "distant", "peppery", "nonstop", "muffled", "pleased", "faraway", "dearest", "orderly", "hurtful", "hateful", "thrifty", "general", "amazing", "upright", "prickly", "visible", "oddball", "opulent", "shadowy", "private", "darling", "vibrant", "lasting", "harmful", "notable", "medical", "vicious", "unhappy", "rundown", "content", "made-up", "elastic", "selfish", "organic", "exalted", "dutiful", "grouchy", "strange", "illegal", "blaring", "boiling", "zealous", "shocked", "perfect", "humming", "dimpled", "playful", "honored", "ancient", "awesome", "velvety", "virtual", "parched", "partial", "teeming", "far-off", "buttery", "awkward", "classic", "elegant", "useless", "serious", "hideous", "average", "married", "unkempt", "tedious", "weighty", "worldly", "similar", "obvious", "offbeat", "bruised", "defiant", "hopeful", "ethical", "starchy", "earnest", "buoyant", "squeaky", "uniform", "worried", "digital", "crowded", "corrupt", "trivial", "naughty", "capital", "melodic", "willing", "beloved", "another", "winding", "monthly", "primary", "distant", "flowery", "spanish"],
+                                    10u8 => vec!["delectable", "attractive", "overcooked", "frightened", "disfigured", "unfinished", "well-to-do", "definitive", "reflecting", "easy-going", "impressive", "determined", "unpleasant", "tremendous", "reasonable", "sweltering", "glistening", "remarkable", "optimistic", "motionless", "noteworthy", "disastrous", "remorseful", "altruistic", "immaterial", "authorized", "legitimate", "respectful", "profitable", "impossible", "irritating", "bite-sized", "glittering", "scientific", "avaricious", "supportive", "worthwhile", "submissive", "victorious", "acceptable", "thoughtful", "thunderous", "improbable", "extraneous", "overlooked", "short-term", "passionate", "flamboyant", "enchanting", "elementary", "jam-packed", "high-level", "villainous", "suspicious", "beneficial", "illiterate", "negligible", "nutritious", "delightful", "forthright", "monumental", "occasional", "simplistic", "stupendous", "harmonious", "outrageous", "tremendous", "courageous", "threadbare", "burdensome", "cultivated", "cumbersome", "elliptical", "triangular", "dependable", "mysterious", "shimmering", "outlandish", "infatuated", "adolescent", "gregarious", "incomplete", "idealistic", "disgusting", "flickering", "productive", "immaculate", "functional", "incredible", "impeccable", "voluminous", "hospitable", "equatorial", "verifiable", "serpentine", "aggressive", "gargantuan", "celebrated"],
+                                    15u8 => vec!["well-documented", "black-and-white", "inconsequential"]
+                                    }
+                                    ,GrammaticalClass::Adverbs => map! {
+                                    7u8 => vec!["happily", "kookily", "sharply", "sweetly", "tensely", "briskly", "heavily", "readily", "loyally", "angrily", "lithely", "vaguely", "orderly", "noisily", "clearly", "tightly", "luckily", "rightly", "zestily", "proudly", "nastily", "wearily", "bravely", "plainly", "tepidly", "crossly", "sternly", "harshly", "shrilly", "monthly", "lightly", "shortly", "nightly", "merrily", "fatally", "thereby", "chiefly", "roughly", "densely", "finally", "bleakly", "hastily", "awfully", "visibly", "closely", "wrongly", "greatly", "blindly", "acutely", "usually", "scarily", "utterly", "loosely", "loftily", "cleanly", "already", "rapidly", "frankly", "quietly", "needily", "handily", "morally", "quickly", "equally", "swiftly", "solidly", "rigidly", "exactly", "crazily", "briefly", "shakily", "cruelly", "totally", "eagerly"],
+                                    8u8 => vec!["hungrily", "covertly", "knottily", "reliably", "properly", "normally", "visually", "casually", "silently", "dreamily", "possibly", "politely", "probably", "fiercely", "steadily", "docilely", "joyfully", "cleverly", "boyishly", "stupidly", "queasily", "intently", "directly", "cowardly", "lovingly", "famously", "urgently", "bitterly", "securely", "mortally", "jauntily", "markedly", "slightly", "clumsily", "daintily", "honestly", "actually", "cloudily", "daringly", "suitably", "recently", "upwardly", "smoothly", "vacantly", "speedily", "sedately", "scarcely", "woefully", "strictly", "fluently", "quaintly", "wickedly", "terribly", "formerly", "solemnly", "jaggedly", "entirely", "sleepily", "actively", "jovially", "prettily", "joyously", "worthily", "formally", "tenderly", "suddenly", "verbally", "candidly", "promptly", "usefully", "annually", "commonly", "finitely", "randomly", "heartily", "decently", "brightly", "unkindly", "brutally", "mightily", "quirkily", "greedily", "inwardly", "shoddily"],
+                                    10u8 => vec!["splendidly", "measurably", "confusedly", "handsomely", "gracefully", "thoroughly", "ultimately", "diligently", "positively", "vigorously", "selflessly", "blissfully", "humorously", "constantly", "admiringly", "poetically", "eventually", "officially", "obediently", "singularly", "materially", "touchingly", "externally", "habitually", "abnormally", "vertically", "explicitly", "poignantly", "diagonally", "negatively", "carelessly", "absolutely", "informally", "presumably", "stealthily", "generously", "divisively", "ruthlessly", "recklessly", "hopelessly", "gratefully", "ironically", "innocently", "repeatedly", "hauntingly", "believably", "reasonably", "invariably", "painlessly", "powerfully", "frequently", "especially", "separately", "physically", "arrogantly", "obligingly", "skillfully", "previously", "vigilantly", "abundantly", "youthfully", "internally", "graciously", "ordinarily", "charmingly", "annoyingly", "inevitably", "reponsibly", "studiously", "enormously", "faithfully", "doubtfully", "thankfully", "colorfully", "yearningly", "rightfully", "menacingly", "truthfully", "boastfully", "remarkably", "punctually", "delicately", "eloquently", "childishly", "jubilantly", "completely", "sheepishly", "infinitely", "puzzlingly", "cautiously", "personally", "disorderly", "cheerfully", "grudgingly", "unbearably"],
+                                    11u8 => vec!["obnoxiously", "idiotically", "haphazardly", "resentfully", "unanimously", "righteously", "unfavorably", "offensively", "temporarily", "universally", "voluntarily", "luxuriously", "startlingly", "principally", "fashionably", "wonderfully", "customarily", "emotionally", "ferociously", "extensively", "indubitably", "comfortably", "respectably", "brilliantly", "exclusively", "exceedingly", "elaborately", "immediately", "incessantly", "secretively", "evocatively", "unselfishly", "accordingly", "stressfully", "beautifully", "impatiently", "quizzically", "unnaturally", "shamelessly", "appreciably", "financially", "searchingly", "incorrectly", "unwillingly", "spiritually", "irrefutably", "unfailingly", "dangerously", "potentially", "essentially", "differently", "reluctantly", "fortunately", "strenuously", "vivaciously", "perpetually", "efficiently", "frantically", "unabashedly", "upliftingly", "unethically", "continually"],
+                                    15u8 => vec!["inappropriately", "extraordinarily", "overconfidently", "sympathetically", "compassionately", "whole-heartedly"],
+                                    4u8 => vec!["only", "very", "well"],
+                                    14u8 => vec!["intermittently", "optimistically", "scientifically", "unquestionably", "affectionately", "coincidentally", "absentmindedly", "systematically", "half-heartedly", "unimpressively"],
+                                    13u8 => vec!["knowledgeably", "spectacularly", "spasmodically", "substantially", "imaginatively", "independently", "affirmatively", "exceptionally", "unnecessarily", "significantly", "bewilderingly", "paradoxically", "inquisitively", "considerately", "unemotionally", "silicitiously", "kindheartedly", "unfortunately", "energetically", "appropriately", "reproachfully", "astonishingly", "interestingly", "questioningly", "intelligently", "superficially"],
+                                    9u8 => vec!["selfishly", "viciously", "intensely", "restfully", "partially", "correctly", "expressly", "sincerely", "playfully", "favorably", "ethically", "agreeably", "currently", "naughtily", "earnestly", "elegantly", "painfully", "furiously", "worriedly", "excitedly", "eternally", "generally", "valiantly", "magically", "carefully", "perfectly", "cunningly", "miserably", "medically", "endlessly", "zestfully", "routinely", "patiently", "curiously", "uselessly", "instantly", "enjoyably", "fervently", "genuinely", "specially", "mockingly", "immensely", "dastardly", "healthily", "evidently", "jealously", "longingly", "supremely", "compactly", "hurriedly", "outwardly", "stylishly", "naturally", "foolishly", "seemingly", "bashfully", "defiantly", "irritably", "gleefully", "awkwardly", "extremely", "assuredly", "obviously", "presently", "violently", "willingly", "girlishly", "helpfully", "yesterday", "nervously", "seriously", "hatefully", "willfully", "invisibly", "knavishly", "amazingly", "ominously", "regularly", "gradually", "kiddingly", "coaxingly", "certainly", "uniformly", "zealously", "anxiously", "joylessly", "knowingly"],
+                                    16u8 => vec!["enthusiastically"],
+                                    12u8 => vec!["methodically", "hysterically", "sporadically", "specifically", "majestically", "victoriously", "deliberately", "structurally", "mysteriously", "unexpectedly", "impressively", "sufficiently", "respectfully", "reassuringly", "unskillfully", "periodically", "successfully", "outrageously", "surprisingly", "unbelievably", "particularly", "persistently", "effortlessly", "tremendously", "immeasurably", "infrequently", "accidentally", "bewilderedly", "frenetically", "ridiculously", "suspiciously", "meaningfully", "courageously", "mechanically", "consequently", "persuasively", "passionately", "continuously", "considerably", "horizontally", "subsequently", "thoughtfully", "questionably", "begrudgingly", "judgmentally", "consistently", "delightfully", "harmoniously", "dramatically", "unilaterally", "occasionally", "identifiably"],
+                                    5u8 => vec!["badly", "daily", "fully", "sadly", "fiery", "icily", "slyly", "every", "dimly", "today", "wetly", "oddly", "still", "early", "madly", "shyly", "truly"],
+                                    3u8 => vec!["why"],
+                                    6u8 => vec!["highly", "solely", "dearly", "likely", "lastly", "safely", "wisely", "widely", "justly", "surely", "warmly", "yearly", "weekly", "barely", "primly", "freely", "coolly", "poorly", "vastly", "fondly", "merely", "lately", "nearly", "wildly", "lazily", "simply", "boldly", "mainly", "limply", "rudely", "easily", "softly", "mostly", "busily", "rarely", "anyway", "evenly", "dopily", "richly", "openly", "deftly", "grimly", "keenly", "subtly", "meanly", "loudly", "wiggly", "weakly", "lively", "zanily", "calmly", "nosily", "nicely", "vainly", "gently", "coldly", "nearby", "darkly", "slowly", "wholly", "numbly", "neatly", "really", "hourly", "gladly", "deeply", "kindly", "fairly", "hugely", "cutely"]
+                                    }
+                                    ,GrammaticalClass::PluralNouns => map! {
+                                    11u8 => vec!["comparisons", "cassowaries", "toothpastes", "salamanders", "polar bears", "plantations", "goldfinches", "acousticses", "volleyballs", "playgrounds", "experiences", "fellowships", "butterflies", "chinchillas", "arithmetics", "suggestions", "earthquakes", "basketballs", "calculators", "sea urchins", "punishments", "secretaries", "albatrosses", "attractions", "connections", "apparatuses", "jellyfishes", "discussions", "instruments", "adjustments", "guinea pigs", "authorities", "territories", "compatriots", "chimpanzees", "governments", "dragonflies", "cockroaches", "discoveries", "cab drivers", "cartoonists", "contractors", "salespeople", "ambassadors", "astronomers", "auctioneers", "undertakers"],
+                                    8u8 => vec!["termites", "thoughts", "produces", "popcorns", "manatees", "dolphins", "chickens", "vicuñas", "requests", "panthers", "opinions", "feelings", "lookouts", "walruses", "controls", "payments", "potatoes", "insurers", "quartzes", "opossums", "partners", "contacts", "raccoons", "pelicans", "sweaters", "cabbages", "aviators", "stitches", "lapwings", "oatmeals", "memories", "narwhals", "readings", "recesses", "officers", "hydrants", "diseases", "examples", "hamsters", "airports", "students", "koupreys", "quarters", "visitors", "disgusts", "canvases", "purposes", "haircuts", "churches", "creators", "drivings", "gorillas", "whistles", "attempts", "cactuses", "goshawks", "lettuces", "managers", "curtains", "caribous", "polishes", "cleaners", "pictures", "railways", "gazelles", "busheses", "fictions", "tarsiers", "holidays", "accounts", "achieves", "echidnas", "mammoths", "currents", "sardines", "thunders", "pancakes", "brothers", "theories", "cherries", "respects", "impulses", "mornings", "peafowls", "giraffes", "leopards", "captions", "troubles", "balances", "shooters", "lobsters", "apparels", "balloons", "comforts", "cushions", "penguins", "hearings", "meetings", "bedrooms", "branches", "flaggers", "fighters", "crackers", "feathers", "artisans", "stations", "sciences", "measures", "channels", "receipts", "protests", "writings", "supports", "soldiers", "celeries", "wrenches", "cheetahs", "textures", "herrings", "journeys", "servants", "cowhands", "laborers", "machines", "mallards", "believes", "scholars", "weathers", "leathers", "plastics", "insiders", "ladybugs", "quetzals", "emperors", "captains", "cashiers", "catchers", "caterers", "cellists", "chemists", "cobblers", "coroners", "couriers", "samurais", "senators", "sheriffs", "sweepers", "surgeons", "swimmers", "veterans", "athletes"],
+                                    4u8 => vec!["gods", "yaks", "cubs", "nuts", "pens", "ices", "nets", "cans", "arts", "rays", "cods", "zoos", "ants", "gnus", "caps", "pigs", "rubs", "rats", "vans", "jars", "hots", "emus", "arms", "fans", "keys", "dogs", "guns", "lips", "apes", "beds", "rams", "pots", "ears", "cups", "legs", "wars", "rods", "runs", "airs", "eggs", "maps", "inks", "days", "pets", "eels", "eyes", "uses", "toys", "acts", "tubs", "seas", "tops", "hats", "tins", "fogs", "jams", "lows", "elks", "bats", "owls", "jays", "cars", "pans", "hens", "suns", "pies", "ways", "cats", "pins", "yams", "toes", "buns", "bees", "oils", "bits", "ends", "ones", "cops"],
+                                    13u8 => vec!["choir members", "distributions", "faith healers", "superstitions", "water buffalo", "stegosauruses", "afterthoughts", "sharpshooters", "organizations", "entomologists", "entrepreneurs", "exterminators", "calligraphers", "cardiologists", "cartographers", "mail carriers", "manufacturers", "statisticians", "veterinarians"],
+                                    15u8 => vec!["representatives", "press operators", "security guards"],
+                                    14u8 => vec!["komodo dragons", "brew directors", "spirit healers", "chorus singers", "businesspeople", "postal workers", "elephant seals", "advertisements", "cryptographers", "mathematicians"],
+                                    10u8 => vec!["structures", "alligators", "boundaries", "deliverers", "afternoons", "passengers", "appliances", "conductors", "assistants", "businesses", "agreements", "industries", "cemeteries", "englanders", "inventions", "mosquitoes", "doorkeepes", "breakfasts", "goosanders", "trouserses", "rainstorms", "homemakers", "astronauts", "treatments", "underwears", "existences", "activities", "carpenters", "sandpipers", "operations", "armadillos", "red pandas", "selections", "tendencies", "expansions", "furnitures", "cormorants", "transports", "substances", "porcupines", "directions", "quicksands", "cherrieses", "aftermaths", "insurances", "scapegoats", "amusements", "crocodiles", "attentions", "partridges", "go-getters", "digestions", "wolverines", "enchanters", "lunchrooms", "goldfishes", "educations", "barracudas", "scissorses", "vegetables", "knowledges", "conditions", "properties", "scarecrows", "committees", "statements", "principals", "pollutions", "ecologists", "economists", "executives", "caregivers", "chauffeurs", "concierges", "custodians", "messengers", "scientists", "violinists", "librarians", "zookeepers", "zoologists"],
+                                    18u8 => vec!["radio broadcasters"],
+                                    7u8 => vec!["porters", "battles", "flights", "curlews", "galleys", "drawers", "rabbits", "buffalo", "sonsuns", "bubbles", "friends", "sisters", "cattles", "tongues", "weights", "cheeses", "glasses", "pencils", "pigeons", "finches", "sponges", "dinners", "attacks", "fleshes", "bottles", "parrots", "coyotes", "waveses", "wealths", "sneezes", "crushes", "crayons", "eggnogs", "lockets", "riddles", "jeanses", "classes", "cobwebs", "lunches", "queleas", "kitties", "baskets", "turkeys", "engines", "threads", "effects", "squares", "vessels", "streets", "silvers", "icicles", "profits", "mittens", "incomes", "minutes", "marbles", "credits", "chesses", "buttons", "nations", "motions", "oysters", "parties", "amounts", "cameras", "poisona", "throats", "experts", "quinces", "sailors", "zippers", "bananas", "anchors", "thrills", "watches", "ferrets", "lawyers", "magpies", "weasels", "fingers", "flowers", "salmons", "clovers", "harbors", "gerbils", "chances", "falcons", "pockets", "patches", "summers", "kettles", "degrees", "reasons", "turtles", "alpacas", "brushes", "doctors", "tempers", "breaths", "faucets", "systems", "beetles", "advices", "tickets", "matches", "locusts", "brasses", "clerics", "letters", "healths", "answers", "borders", "beavers", "hornets", "voyages", "gardens", "schools", "parcels", "muscles", "cellars", "guitars", "oranges", "runners", "middles", "baboons", "jellies", "thrones", "coaches", "singers", "prisons", "hammers", "monkeys", "pickles", "cannons", "dresses", "numbers", "farmers", "choughs", "grouses", "beggars", "ploughs", "flavors", "animals", "grasses", "powders", "lumbers", "needles", "changes", "stories", "wombats", "jaguars", "dunlins", "kittens", "spiders", "parents", "beliefs", "designs", "berries", "insects", "actions", "rhythms", "coppers", "streams", "offices", "windows", "details", "markets", "zephyrs", "records", "winters", "circles", "regrets", "priests", "butters", "badgers", "dugongs", "lorises", "jackals", "springs", "polices", "collars", "quivers", "buckets", "writers", "strings", "liquids", "islands", "rewards", "touches", "bridges", "growths", "grounds", "editors", "consuls", "marines", "streets", "archers", "artists", "authors"],
+                                    5u8 => vec!["pumps", "clubs", "swans", "units", "boxes", "fires", "yokes", "trips", "waxes", "beefs", "beads", "casts", "ducks", "mices", "irons", "sands", "taxes", "rails", "cribs", "rocks", "soaps", "sides", "birds", "hopes", "drugs", "looks", "snows", "sacks", "talks", "ideas", "cards", "lamps", "gaurs", "loves", "goats", "swims", "locks", "drops", "combs", "nests", "wines", "golds", "drums", "pipes", "meals", "debts", "slips", "leads", "backs", "mines", "heats", "hares", "pails", "skins", "songs", "pests", "coils", "stews", "dirts", "faces", "signs", "homes", "chins", "sizes", "yarns", "woods", "heirs", "bulbs", "frogs", "whips", "facts", "knees", "moles", "moves", "roots", "beans", "dusts", "selfs", "bases", "holes", "moons", "winds", "newts", "glues", "kudus", "games", "rooms", "bears", "seeds", "falls", "masks", "fears", "jokes", "views", "tents", "kites", "males", "trays", "wools", "folds", "vases", "boots", "plays", "spies", "sofas", "waves", "hills", "works", "heads", "pains", "teams", "bombs", "lists", "words", "lions", "twigs", "fangs", "cooks", "beams", "edges", "halls", "carts", "knots", "soups", "crabs", "stems", "mints", "sodas", "nails", "sinks", "minks", "names", "ships", "laces", "vests", "bites", "hours", "plots", "wings", "socks", "hoses", "zincs", "walls", "veils", "coals", "towns", "cakes", "stops", "loafs", "deers", "skies", "hooks", "wires", "burns", "joins", "forms", "seats", "stars", "rooks", "turns", "maids", "wrens", "lakes", "bones", "steps", "rules", "mists", "lines", "desks", "rests", "weeks", "corns", "gulls", "roses", "shoes", "rices", "doors", "roads", "cords", "jails", "hates", "feets", "dolls", "pears", "dimes", "corks", "salts", "foots", "doves", "books", "foods", "crows", "spots", "silks", "trees", "cents", "sails", "lands", "times", "noses", "gnats", "rates", "jumps", "rings", "pulls", "baths", "forks", "larks", "ovens", "balls", "parks", "cries", "tails", "docks", "walks", "helps", "suits", "needs", "camps", "rains", "shops", "veins", "fuels", "marks", "clams", "parts", "horns", "bikes", "coats", "boats", "boars", "rakes", "milks", "toads", "wolfs", "minds", "meats", "worms", "leafs", "shows", "ables", "years", "lifts", "baits", "tanks", "gates", "farms", "roofs", "seals", "bells", "necks", "kicks", "flags", "pages", "yards", "hawks", "fowls", "tests", "wasps", "flies", "notes", "cares", "bands", "hairs", "sorts", "rolls", "foxes", "caves", "chefs", "monks"],
+                                    6u8 => vec!["thumbs", "actors", "prints", "tigers", "sharks", "divers", "quiets", "titles", "apples", "metals", "tastes", "quails", "smiles", "crates", "slopes", "papers", "wastes", "fifths", "quills", "spoons", "humors", "musics", "stamps", "honeys", "bodies", "peaces", "drinks", "ponies", "lemurs", "crimes", "fights", "trails", "ushers", "feasts", "scarfs", "smokes", "otters", "nights", "sleeps", "linens", "scenes", "robots", "chairs", "frames", "wounds", "magics", "sloths", "rivers", "errors", "colors", "gooses", "crooks", "houses", "horses", "pushes", "laughs", "sparks", "geeses", "wishes", "screws", "groups", "crowds", "cables", "straws", "prices", "wheels", "tramps", "knifes", "babies", "cracks", "teeths", "grades", "orders", "waters", "senses", "crowns", "people", "births", "shapes", "koalas", "brains", "causes", "fronts", "covers", "bursts", "scents", "planes", "plates", "ranges", "paints", "deaths", "guides", "floors", "tables", "okapis", "blades", "stones", "whales", "powers", "places", "routes", "steels", "trucks", "months", "sleets", "drains", "hippos", "trains", "flocks", "jurors", "limits", "coatis", "eagles", "boards", "cobras", "shakes", "masses", "hearts", "events", "copies", "skates", "voices", "fishes", "spaces", "spades", "norths", "fields", "skunks", "snails", "gloves", "juices", "pandas", "washes", "moneys", "basins", "llamas", "lights", "plants", "things", "clocks", "ibexes", "fruits", "sheets", "angers", "slaves", "stages", "owners", "mooses", "zebras", "starts", "cloths", "nerves", "beasts", "louses", "bisons", "noises", "clouds", "mouths", "coasts", "breads", "camels", "heroes", "tapirs", "creams", "hyenas", "bricks", "fixers", "shames", "arches", "coughs", "shelfs", "chalks", "earths", "bosses", "elands", "shades", "shocks", "tricks", "alleys", "stores", "staffs", "shrews", "skirts", "herons", "levels", "flames", "armies", "shirts", "stoves", "bloods", "sounds", "forces", "scales", "robins", "wrists", "steams", "chains", "mouses", "points", "ravens", "proses", "losses", "twists", "ghosts", "newses", "quilts", "snakes", "sugars", "tooths", "badges", "brakes", "judges", "smells", "queens", "verses", "trades", "rifles", "cranes", "sheeps", "grapes", "elbows", "curves", "grains", "swings", "jewels", "values", "alarms", "pastes", "ibises", "offers", "angles", "oceans", "clerks", "masons", "mayors", "miners", "models", "scouts", "smiths", "valets", "vicars"],
+                                    12u8 => vec!["toothbrushes", "observations", "housekeepers", "floorwalkers", "developments", "spokespeople", "guinea fowls", "caterpillars", "nightingales", "firefighters", "prairie dogs", "hummingbirds", "competitions", "flesh eaters", "grasshoppers", "destructions", "wildernesses", "sand dollars", "giant pandas", "electricians", "entertainers", "saxophonists", "scuba divers", "stockbrokers", "xylophonists"],
+                                    17u8 => vec!["flight attendants"],
+                                    9u8 => vec!["strangers", "questions", "dogfishes", "mandrills", "processes", "learnings", "good-byes", "kinkajous", "foresters", "creatures", "calendars", "dinosaurs", "starlings", "arguments", "libraries", "relations", "birthdays", "frictions", "commoners", "surprises", "ancestors", "downtowns", "mailboxes", "ostriches", "lyrebirds", "sorcerers", "histories", "mountains", "airplanes", "reporters", "vacations", "achievers", "stretches", "comedians", "governors", "increases", "notebooks", "squirrels", "newcomers", "buildings", "seahorses", "behaviors", "religions", "harmonies", "approvals", "octopuses", "carriages", "teachings", "umbrellas", "sculptors", "hedgehogs", "decisions", "graduates", "aardvarks", "volcanoes", "beginners", "anteaters", "porpoises", "additions", "pleasures", "enlistees", "countries", "divisions", "positions", "ornaments", "mongooses", "flamingos", "wallabies", "stomaches", "companies", "antelopes", "stockings", "kangaroos", "seashores", "hobbieses", "reactions", "ministers", "sidewalks", "exchanges", "distances", "sea lions", "pheasants", "dotterels", "societies", "reindeers", "red deers", "hospitals", "daughters", "interests", "baseballs", "languages", "elephants", "educators", "empresses", "engineers", "explorers", "exporters", "chaplains", "composers", "magicians", "marketers", "mechanics", "muralists", "musicians", "surveyors", "animators", "attorneys"]
+                                    }
+                    };
+        };
+
+        ///The source text grouped by (word, next word) [following words]
+        pub static ref MARKOV_CHAIN: HashMap<(&'static str, &'static str), Vec<&'static str>> = {
+                println!("creeeeating ref");
+                let source = include!("data/source-text");
+                let split_words: Vec<&str> = source.split_ascii_whitespace().collect();
+                let mut markov_chain: HashMap<(&str, &str), Vec<&str>> = HashMap::new();
+
+                for i in 0..(split_words.len() - 3) {
+                        let key = (split_words[i], split_words[i + 1]);
+                        let value = split_words[i + 2];
+                        match markov_chain.get_mut(&key) {
+                                Some(words) => {
+                                        words.push(value);
+                                }
+                                None => {
+                                        markov_chain.insert(key, vec![value]);
+                                }
+                        }
+                }
+                return markov_chain;
+        };
+}
