+++ native/src/name.rs
@@ -1,52 +1,13 @@
-extern crate neon;
-extern crate strum;
-#[macro_use]
-extern crate strum_macros;
-extern crate either;
-extern crate num_derive;
-extern crate num_traits;
-extern crate rand;
-extern crate unicode_segmentation;
-
-use neon::prelude::*;
-use rand::Rng;
-use std::collections::HashMap;
-mod data;
-mod types;
-use data::populate_grammatical_classes;
+use data::GRAMATICAL_CLASSES;
 use either::Either;
 use either::Either::*;
-use num_traits::FromPrimitive;
+use rand::Rng;
 use types::*;
-use unicode_segmentation::UnicodeSegmentation;
-
-fn generate(mut cx: FunctionContext) -> JsResult<JsString> {
-        let what = cx.argument::<JsNumber>(0)?.value();
-        let chars = cx.argument::<JsNumber>(1)?.value() as u8;
-        let result = match FromPrimitive::from_f64(what) {
-                Some(What::Name) => name(chars),
-                Some(What::Description) => name(chars),
-                None => panic!("pattern matching failed for generate"),
-        };
-
-        return Ok(cx.string(capitalize(result)));
-}
-
-register_module!(mut cx, { cx.export_function("generate", generate) });
-
-fn capitalize(text: String) -> String {
-        let mut letters = text.chars();
-
-        match letters.next() {
-                None => String::new(),
-                Some(l) => l.to_uppercase().chain(letters).collect(),
-        }
-}
 
 /// Names are generated according to the following patterns:
 ///  <adjective> [, <adjective>] <noun>
 ///  [<adjective>] <noun> [who | that] [<adverb>] <verb> [<adjective>] [<noun>] [<other>]
-fn name(total_chars: u8) -> String {
+pub fn generate_name(total_chars: u8) -> String {
         if should_happen(60) {
                 return simple_name(total_chars);
         }
@@ -67,7 +28,6 @@ fn simple_name(remaining_chars: u8) -> String {
                 return format!("{}, {} {}", names[0], names[1], names[2]);
         }
         return names.join(" ");
-
 }
 
 /// A "complex name" is a string containing [<adjective>] <noun> [who | that] [<adverb>] <verb> [<adjective>] [<noun>] [<other>]
@@ -81,32 +41,30 @@ fn complex_name(remaining_chars: u8) -> String {
                 (Right(GrammaticalClass::Adjectives), 10),
                 (Right(GrammaticalClass::PluralNouns), 30),
                 (Right(GrammaticalClass::Other), 5),
-        ] ;
+        ];
         return make_names(remaining_chars, classes_chances).join(" ");
 }
 
-fn make_names(mut remaining_chars: u8, classes_chances: Vec<(Either<&str, GrammaticalClass>, u8)>) -> Vec<String> {
-        let grammatical_classes = populate_grammatical_classes();
+fn make_names(
+        mut remaining_chars: u8,
+        classes_chances: Vec<(Either<&str, GrammaticalClass>, u8)>,
+) -> Vec<String> {
         let mut names: Vec<String> = Vec::new();
 
         for class in filter_by_chance(classes_chances) {
                 match class {
                         Right(cl) => {
-                                let word = get_word(&grammatical_classes, cl, remaining_chars);
+                                let word = get_word(cl, remaining_chars);
 
-                                match word {
-                                        Some(wd) => {
-                                                remaining_chars -= wd.len() as u8;
-                                                names.push(wd);
-                                        }
-                                        None => {}
+                                if let Some(wd) = word {
+                                        remaining_chars -= wd.len() as u8;
+                                        names.push(wd);
                                 }
                         }
                         Left(word) => {
-                                let size = word.graphemes(true).count() as u8;
+                                let size = word.len() as u8;
 
-                                if remaining_chars - 1 >= size
-                                {
+                                if remaining_chars - 1 >= size {
                                         names.push(word.to_string());
                                         remaining_chars -= size;
                                 }
@@ -119,18 +77,17 @@ fn make_names(mut remaining_chars: u8, classes_chances: Vec<(Either<&str, Gramma
 
 /// Pick a gramatical class random word up to the given size
 fn get_word(
-        grammatical_classes: &HashMap<GrammaticalClass, HashMap<u8, Vec<&'static str>>>,
         class: GrammaticalClass,
         remaining_chars: u8,
 ) -> Option<String> {
-        let keys: Vec<u8> = grammatical_classes[&class]
+        let keys: Vec<u8> = GRAMATICAL_CLASSES[&class]
                 .keys()
                 .filter(|&k| *k <= remaining_chars - 1)
                 .cloned()
                 .collect();
 
         if keys.len() > 0 {
-                let words = &grammatical_classes[&class]
+                let words = &GRAMATICAL_CLASSES[&class]
                         [&keys[rand::thread_rng().gen_range(0, keys.len())]];
                 let word = &words[rand::thread_rng().gen_range(0, words.len())];
 
@@ -141,17 +98,20 @@ fn get_word(
 }
 
 /// Given a list, returns values according to their chance of occuring
-fn filter_by_chance(classes: Vec<(Either<&str, GrammaticalClass>, u8)>) -> Vec<Either<&str, GrammaticalClass>> {
-        classes.iter()
+fn filter_by_chance(
+        classes: Vec<(Either<&str, GrammaticalClass>, u8)>,
+) -> Vec<Either<&str, GrammaticalClass>> {
+        return classes
+                .iter()
                 .filter(|(_, chance)| should_happen(*chance))
                 .map(|(class, _)| class)
                 .cloned()
-                .collect()
+                .collect();
 }
 
 /// Chance of an event occuring out of 100 times
 fn should_happen(chance: u8) -> bool {
-        rand::thread_rng().gen_range(1, 100) <= chance
+        return rand::thread_rng().gen_range(1, 100) <= chance;
 }
 
 #[cfg(test)]
@@ -160,25 +120,19 @@ mod tests {
 
         #[test]
         fn should_happen_test() {
-                assert_eq!(should_happen(100), true);
-                assert_eq!(should_happen(0), false);
+                assert_eq!(true, should_happen(100));
+                assert_eq!(false, should_happen(0));
         }
 
         #[test]
         fn filter_by_chance_test() {
                 assert_eq!(
-                        filter_by_chance(vec![(Right(GrammaticalClass::Adjectives), 100u8)]),
-                        vec![Right(GrammaticalClass::Adjectives)]
+                        vec![Right(GrammaticalClass::Adjectives)],
+                        filter_by_chance(vec![(Right(GrammaticalClass::Adjectives), 100u8)])
                 );
-                assert_eq!(
-                        filter_by_chance(vec![(Right(GrammaticalClass::Adjectives), 0u8)]),
-                        vec![]
+                assert!(
+                        filter_by_chance(vec![(Right(GrammaticalClass::Adjectives), 0u8)])
+                                .is_empty()
                 );
         }
-
-        //for debugging purposes
-        // #[test]
-        // fn simple_name_test() {
-        //         populate_grammatical_classes();
-        // }
 }
