--- rs/.gitignore
@@ -1,6 +1,0 @@
-native/target
-native/index.node
-native/artifacts.json
-**/*~
-**/node_modules
-**/.DS_Store
