--- rs/lib/index.js
@@ -1,3 +1,0 @@
-var addon = require('../native');
-
-console.log(addon.generate(1, 100));
